20-10-2022 16:47

+++ banking-backend/src/main/java/com/banking/repository/TransferRepository.java
@@ -10,7 +10,7 @@ import com.banking.model.Transfer;
 public interface TransferRepository extends JpaRepository<Transfer, Integer> {
 
 	@Query("select t from Transfer t where t.from_account = ?1")
-	List<Transfer> listByAccountNo(int account_no);
+	List<Transfer> listByAccountNo(long account_no);
